Include explicit resource requests for initContainter
Without these requests, if you have a quota set on the
namespace, the initContainer cannot start and then the
main Jupyter container cannot start either.

For now, just duplicate the resource request for the main
container, but it is likely that we will want to adjust
the resource request for the initContainer to be smaller;
it may even be hard coded since the user does not see the
initContainer.